backend/requirements.txt
# PostForge AI — Social Media Agent
# Multi-tenant AI-powered social media automation SaaS
# © 2025 Mamu — All Rights Reserved

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
openai==1.55.3
facebook-sdk==3.1.0
tweepy==4.14.0
linkedin-api==2.0.2
instagram-auto-publish==0.1.6
apscheduler==3.11.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
boto3==1.35.0
redis==5.1.1
celery==5.4.0
flower==2.0.1
python-dotenv==1.0.1
slowapi==0.1.9
requests==2.32.3
